requirements.txt
fastapi
uvicorn
celery
redis
python-dotenv
google-generativeai
langchain-community
pypdf